requirements.txt
aiohttp==3.11.18
pillow==11.2.1
python-dotenv==1.1.0
qrcode==8.1
